Switch to Zeitwerk for code loading
