+++ lib/literal/flags/flags_16.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Literal::Flags16 < Literal::Flags
+	BYTES = 2
+	BITS = BYTES * 8
+	PACKER = "S"
+end
