+++ lib/literal/types/deferred_type.rb
