+++ literal.gemspec
@@ -27,5 +27,7 @@ Gem::Specification.new do |spec|
 
 	spec.require_paths = ["lib"]
 
+	spec.add_runtime_dependency "zeitwerk"
+
 	spec.metadata["rubygems_mfa_required"] = "true"
 end
